Remove run command in favor of better cli usability
