+++ cmd/cmd_root.go
@@ -3,16 +3,20 @@ package cmd
 import (
 	"actionforge/graph-runner/core"
 	"actionforge/graph-runner/utils"
+	u "actionforge/graph-runner/utils"
 	"flag"
 	"fmt"
 	"os"
 	"strings"
 
 	"github.com/spf13/cobra"
+
+	// initialize all nodes
+	_ "actionforge/graph-runner/nodes"
 )
 
 var cmdRoot = &cobra.Command{
-	Use:     "graph-runner",
+	Use:     "graph-runner [filename]",
 	Short:   "Graph runner is a tool for running action graphs.",
 	Version: core.GetFulllVersionInfo(),
 	Run: func(cmd *cobra.Command, args []string) {
@@ -23,19 +27,47 @@ var cmdRoot = &cobra.Command{
 func Execute() {
 	_ = cmdRoot.Flags().Parse(os.Args[1:])
 
-	var cmd = cmdRoot
-
-	// default cmd if no cmd is given
-	if len(os.Args) > 1 && strings.HasSuffix(os.Args[1], ".yml") {
-		args := append([]string{"run"}, os.Args[1:]...)
-		cmd.SetArgs(args)
+	var filename string
+	if len(os.Args) > 1 {
+		if strings.HasSuffix(os.Args[1], ".yml") {
+			filename = os.Args[1]
+		}
 	}
 
-	err := cmd.Execute()
+	if filename == "" {
+		filename = u.GetVariable("graph_file", "The graph file to use", u.GetVariableOpts{
+			Env: true,
+		})
+	}
+
+	if filename != "" {
+		err := ExecuteRun(filename)
+		if err != nil {
+			fmt.Println(err)
+			os.Exit(1)
+		}
+	} else {
+		err := cmdRoot.Execute()
+		if err != nil {
+			fmt.Println(err)
+			os.Exit(1)
+		}
+	}
+}
+
+func ExecuteRun(graphFile string) error {
+
+	graphContent, err := os.ReadFile(graphFile)
 	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
+		return err
 	}
+
+	err = core.RunGraph(graphContent)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
 
 func init() {
